% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_diverging.R
\name{geom_bar_diverging}
\alias{geom_bar_diverging}
\alias{geom_bar_likert}
\alias{geom_bar_pyramid}
\title{Geometries for diverging bar plots}
\usage{
geom_bar_diverging(
  mapping = NULL,
  data = NULL,
  stat = "prop",
  position = position_diverging(reverse = reverse, exclude_fill_values =
    exclude_fill_values, cutoff = cutoff),
  ...,
  complete = "fill",
  default_by = "total",
  vjust = 1,
  reverse = FALSE,
  exclude_fill_values = NULL,
  cutoff = NULL
)

geom_bar_likert(
  mapping = NULL,
  data = NULL,
  stat = "prop",
  position = position_likert(reverse = reverse, exclude_fill_values =
    exclude_fill_values, cutoff = cutoff),
  ...,
  complete = "fill",
  default_by = "x",
  vjust = 1,
  reverse = FALSE,
  exclude_fill_values = NULL,
  cutoff = NULL
)

geom_bar_pyramid(
  mapping = NULL,
  data = NULL,
  stat = "prop2",
  position = position_diverging(reverse = reverse, exclude_fill_values =
    exclude_fill_values, cutoff = cutoff),
  ...,
  complete = "fill",
  default_by = "fill",
  vjust = 1,
  reverse = FALSE,
  exclude_fill_values = NULL,
  cutoff = NULL
)
}
\arguments{
\item{mapping}{Optional set of aesthetic mappings.}

\item{data}{The data to be displayed in this layers.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{A position adjustment to use on the data for this layer.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}}

\item{complete}{Name of an aesthetic determining denominators by default,
see \code{\link[=stat_prop]{stat_prop()}}. Passed only if \code{stat = "prop"}.}

\item{vjust}{Vertical adjustment for geoms that have a position
(like points or lines), not a dimension (like bars or areas). Set to
\code{0} to align with the bottom, \code{0.5} for the middle,
and \code{1} (the default) for the top.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}

\item{exclude_fill_values}{Vector of values from the variable associated with
the \code{fill} aesthetic that should not be displayed (but still taken into
account for computing proportions)}

\item{cutoff}{number of categories to be displayed negatively (i.e. on the
left of the x axis or the bottom of the y axis), could be a decimal value:
\code{2} to display negatively the two first categories, \code{2.5} to display
negatively the two first categories and half of the third, \code{2.2} to display
negatively the two first categories and a fifth of the third (see examples).
By default (\code{NULL}), it will be equal to the number of categories divided
by 2, i.e. it will be centered.}

\item{default_aes}{Default aesthetics.}
}
\description{
These geometries are similar to \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}} but provides
different set of default values.
}
\details{
\itemize{
\item \code{geom_bar_diverging()} is designed for stacked diverging bar plots, using
\code{\link[=position_diverging]{position_diverging()}}.
\item \code{geom_bar_diverging()} is designed for Likert-type items. Using
\code{position_likert()} (each bar sums to 100\%).
\item \code{geom_bar_pyramid()} is adapted for population pyramid plots when a factor
with two levels is mapped to the \strong{fill} aesthetics. The proportions are,
here by default, computed separately for each value of the \strong{fill}
aesthetics.
}
}
\examples{
library(ggplot2)
ggplot(diamonds) +
  aes(y = clarity, fill = cut) +
  geom_bar_diverging()

ggplot(diamonds) +
  aes(y = clarity, fill = cut) +
  geom_bar_diverging(cutoff = 1)

ggplot(diamonds) +
  aes(y = clarity, fill = cut) +
  geom_bar_likert()

d <- Titanic |> as.data.frame()

ggplot(d) +
  aes(y = Class, fill = Sex, weight = Freq) +
  geom_bar_diverging()

ggplot(d) +
  aes(y = Class, fill = Sex, weight = Freq) +
  geom_bar_pyramid()
}
