% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcoef-package.R
\docType{package}
\name{ggcoef-package}
\alias{ggcoef}
\alias{ggcoef-package}
\title{ggcoef: Plot Model Coefficients}
\description{
Provides suite of functions to plot regression model coefficients ("forest plots") using `ggplot2`. The suite includes new geometries to add alternating background color to a plot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://larmarange.github.io/ggcoef/}
  \item Report bugs at \url{https://github.com/larmarange/ggcoef/issues}
}

}
\author{
\strong{Maintainer}: Joseph Larmarange \email{joseph@larmarange.net} (\href{https://orcid.org/0000-0001-7097-700X}{ORCID})

}
\keyword{internal}
