% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_number_abs.R
\name{label_number_abs}
\alias{label_number_abs}
\alias{label_percent_abs}
\title{Label absolute values}
\usage{
label_number_abs(..., hide_below = NULL)

label_percent_abs(..., hide_below = NULL)
}
\arguments{
\item{...}{arguments passed to \code{\link[scales:label_number]{scales::label_number()}} or
\code{\link[scales:label_percent]{scales::label_percent()}}}

\item{hide_below}{if provided, values below \code{hide_below} will be masked
(i.e. an empty string \code{""} will be returned)}
}
\value{
A "labelling" function, , i.e. a function that takes a vector and
returns a character vector of same length giving a label for each input value.
}
\description{
Label absolute values
}
\examples{
x <- c(-0.2, -.05, 0, .07, .25, .66)

scales::label_number()(x)
label_number_abs()(x)

scales::label_percent()(x)
label_percent_abs()(x)
label_percent_abs(hide_below = .1)(x)
}
\seealso{
\code{\link[scales:label_number]{scales::label_number()}}, \code{\link[scales:label_percent]{scales::label_percent()}}
}
