% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_any.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number.}
\source{
adapted from \code{plyr}
}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric or date-time (POSIXct) vector to round}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}}
}
\description{
Round to multiple of any number.
}
\examples{
round_any(1.865, accuracy = .25)
}
