% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sum.R
\name{weighted.sum}
\alias{weighted.sum}
\title{Weighted Sum}
\usage{
weighted.sum(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector of values}

\item{w}{a numeric vector of weights}

\item{na.rm}{a logical indicating whether to ignore \code{NA} values}
}
\value{
A numeric vector.
}
\description{
Weighted Sum
}
\examples{
x <- 1:20
w <- runif(20)
weighted.sum(x, w)
}
