% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstats-package.R
\docType{package}
\name{ggstats-package}
\alias{ggstats}
\alias{ggstats-package}
\title{ggstats: Extension to 'ggplot2' for Plotting Stats}
\description{
Provides new statistics, new geometries and new positions for 'ggplot2' and a suite of functions to facilitate the creation of statistical plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://larmarange.github.io/ggstats/}
  \item \url{https://github.com/larmarange/ggstats}
  \item Report bugs at \url{https://github.com/larmarange/ggstats/issues}
}

}
\author{
\strong{Maintainer}: Joseph Larmarange \email{joseph@larmarange.net} (\href{https://orcid.org/0000-0001-7097-700X}{ORCID})

}
\keyword{internal}
