% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_percent_abs.R
\name{label_number_abs}
\alias{label_number_abs}
\alias{label_percent_abs}
\title{Label absolute values}
\usage{
label_number_abs(...)

label_percent_abs(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[scales:label_number]{scales::label_number()}} or
\code{\link[scales:label_percent]{scales::label_percent()}}}
}
\description{
Label absolute values
}
\examples{
x <- c(-0.2, -.05, 0, .25, .66)

scales::label_number()(x)
label_number_abs()(x)

scales::label_percent()(x)
label_percent_abs()(x)

}
\seealso{
\code{\link[scales:label_number]{scales::label_number()}}, \code{\link[scales:label_percent]{scales::label_percent()}}
}
